+++ src/main/java/shopping/application/AuthService.java
@@ -4,8 +4,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shopping.domain.entity.User;
 import shopping.domain.vo.Password;
-import shopping.dto.LoginRequest;
-import shopping.dto.LoginResponse;
+import shopping.dto.request.LoginRequest;
+import shopping.dto.response.LoginResponse;
 import shopping.exception.BadLoginRequestException;
 import shopping.infrastructure.PasswordEncoder;
 import shopping.infrastructure.TokenProvider;
