+++ src/main/java/shopping/application/CartService.java
@@ -7,9 +7,9 @@ import shopping.domain.entity.CartItem;
 import shopping.domain.entity.Product;
 import shopping.domain.entity.User;
 import shopping.domain.vo.Quantity;
-import shopping.dto.CartCreateRequest;
-import shopping.dto.CartResponse;
-import shopping.dto.QuantityUpdateRequest;
+import shopping.dto.request.CartCreateRequest;
+import shopping.dto.request.QuantityUpdateRequest;
+import shopping.dto.response.CartResponse;
 import shopping.exception.ProductNotFoundException;
 import shopping.exception.UserNotFoundException;
 import shopping.exception.UserNotHasCartItemException;
