+++ src/main/java/shopping/dto/request/QuantityUpdateRequest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.request;
 
 import shopping.domain.vo.Quantity;
