+++ src/main/java/shopping/dto/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 public class ErrorResponse {
     private String message;
