+++ src/main/java/shopping/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 public class LoginResponse {
     private String accessToken;
