+++ src/main/java/shopping/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 import shopping.domain.entity.Product;
