+++ src/test/java/shopping/dto/request/CartCreateRequestTest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
