+++ src/test/java/shopping/dto/response/ProductResponseTest.java
@@ -1,4 +1,4 @@
-package shopping.dto;
+package shopping.dto.response;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
