+++ src/test/java/shopping/integration/AuthIntegrationTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import shopping.dto.ErrorResponse;
-import shopping.dto.LoginResponse;
+import shopping.dto.response.ErrorResponse;
+import shopping.dto.response.LoginResponse;
 import shopping.infrastructure.TokenProvider;
 import shopping.integration.config.IntegrationTest;
 import shopping.integration.util.AuthUtil;
